♻️ Refactor: Slider 컴포넌트 삭제 (#30)
